Abstracted Admin Server class
